refactor(whatsapp-simulator): split questionnaire-chat into modules
Move the message/phase types, bot constants and intro helpers to
src/lib/chat.ts, and extract MessageRow and the Status/Notice views into
src/components/chat/. Drop the default create-next-app favicon.
