+++ hl7-ai-competition/whatsapp-simulator/knip.json
@@ -1,5 +1,4 @@
 {
   "$schema": "https://unpkg.com/knip@6/schema.json",
-  "ignore": ["src/components/ui/**"],
-  "ignoreDependencies": ["@biomejs/biome"]
+  "ignore": ["src/components/ui/**"]
 }
