+++ hl7-ai-competition/whatsapp-simulator/src/components/chat/message-row.tsx
@@ -1,0 +1,82 @@
+import type { Message } from "@/lib/chat";
+
+import { Reply } from "lucide-react";
+
+import { BOT_INITIALS, BOT_NAME, truncate } from "@/lib/chat";
+
+export function MessageRow({
+  message,
+  previous,
+  canReply,
+  onReply,
+}: {
+  message: Message;
+  previous?: Message;
+  canReply: boolean;
+  onReply: () => void;
+}) {
+  const me = message.role === "user";
+  const senderChanged = !previous || previous.role !== message.role;
+  const showAvatar = !me && senderChanged;
+
+  return (
+    <div
+      className="group flex items-start gap-[9px]"
+      style={{
+        marginTop: senderChanged ? 12 : 2,
+        flexDirection: me ? "row-reverse" : "row",
+      }}
+    >
+      <div className="w-7 shrink-0">
+        {showAvatar && (
+          <div className="flex size-7 items-center justify-center rounded-full bg-[#e4e4e7] text-[11px] font-semibold text-[#52525b]">
+            {BOT_INITIALS}
+          </div>
+        )}
+      </div>
+      <div
+        className="flex min-w-0 flex-col"
+        style={{ alignItems: me ? "flex-end" : "flex-start" }}
+      >
+        {message.replyTo && (
+          <div
+            className="-mb-1 flex max-w-[min(78vw,440px)] gap-1 overflow-hidden whitespace-nowrap border-l-2 border-[#d4d4d8] px-2.5 pb-2 pt-1 text-xs"
+            style={{ marginLeft: me ? 0 : 4, marginRight: me ? 4 : 0 }}
+          >
+            <span className="font-medium opacity-80">{BOT_NAME}</span>
+            <span className="opacity-[0.55]">
+              {truncate(message.replyTo.questionText, 52)}
+            </span>
+          </div>
+        )}
+        <div
+          className="flex items-center gap-[7px]"
+          style={{ flexDirection: me ? "row-reverse" : "row" }}
+        >
+          <div
+            className={
+              me
+                ? "max-w-[min(78vw,440px)] whitespace-pre-wrap break-words rounded-[13px] rounded-br-[4px] bg-[#0a0a0a] px-3 py-2 text-sm leading-normal text-[#fafafa]"
+                : "max-w-[min(78vw,440px)] whitespace-pre-wrap break-words rounded-[13px] rounded-bl-[4px] bg-[#f4f4f5] px-3 py-2 text-sm leading-normal text-[#18181b]"
+            }
+          >
+            {message.text}
+          </div>
+          {canReply && (
+            <button
+              type="button"
+              onClick={onReply}
+              aria-label="Reply"
+              className="flex size-6 shrink-0 items-center justify-center rounded-md border border-[#ececed] bg-white text-[#8a8a8e] opacity-0 transition hover:bg-[#f7f7f8] hover:text-[#0a0a0a] group-hover:opacity-100"
+            >
+              <Reply className="size-3" />
+            </button>
+          )}
+        </div>
+        <div className="mt-[3px] px-[3px] text-[11px] text-[#a1a1aa]">
+          {message.time}
+        </div>
+      </div>
+    </div>
+  );
+}
